agent_dispute_policy/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
PyYAML==6.0.1